Revert "lex MinInt; store Integer as Long in token"
This reverts commit 5bfac9a10c663e9b22ce175d43b78a509125dedf.

+++ src/main/java/mtm68/ast/nodes/IntLiteral.java
@@ -1,10 +1,12 @@
 package mtm68.ast.nodes;
 
+import java.math.BigInteger;
+
 import edu.cornell.cs.cs4120.util.SExpPrinter;
 
-public class IntLiteral extends Literal<Long> {
+public class IntLiteral extends Literal<BigInteger> {
 	
-	public IntLiteral(Long value) {
+	public IntLiteral(BigInteger value) {
 		super(value);
 	}
